Reorganize code in scrapertest package
The ability to compare metric related structs is very useful
in scraper testing, but it does not need to be the exclusive
purpose of this package. Anticipating the addition of other
capabilities, the files have been renamed to reflect their
specific purposes.

No functional changes are made here.
