+++ internal/scrapertest/compare.go
